Remove filesystem.ts
