+++ src/components/app/exporters/exportersCommon.ts
@@ -1,8 +1,8 @@
-import { cloneDeep } from 'lodash';
-import { computed, reactive, ref } from '@vue/composition-api';
-import { VForm } from '@/types/vuetify';
-import store from '@/store';
 import { OutputVendorName } from '@/originalBudgetTrackingApp/commonTypes';
+import store from '@/store';
+import { VForm } from '@/types/vuetify';
+import { computed, reactive, ref } from '@vue/composition-api';
+import { cloneDeep } from 'lodash';
 
 export function setupExporterConfigForm<T extends OutputVendorName>(exporterName: T) {
   const vForm = ref<VForm>();
