+++ src/originalBudgetTrackingApp/export/outputVendors/googleSheets/googleSheetsInternalAPI.ts
@@ -46,7 +46,7 @@ export async function getExistingHashes({
   return existingHashes;
 }
 
-async function loadCredentialsAndAuthorize(credentialsFilePath: string) {
+async function loadCredentialsAndAuthorize(credentialsFilePath) {
   const credentialsStr = await readFile(credentialsFilePath, 'utf8');
   const credentials = JSON.parse(credentialsStr);
   const jwtClient = new google.auth.JWT(credentials.client_email, undefined, credentials.private_key, SCOPES);
