+++ src/originalBudgetTrackingApp/export/outputVendors/index.ts
@@ -1,6 +1,7 @@
-import { ynabOutputVendor } from './ynab/ynab';
 import { googleSheetsOutputVendor } from './googleSheets/googleSheets';
+import json from './json/json';
+import { ynabOutputVendor } from './ynab/ynab';
 
-const outputVendors = [ynabOutputVendor, googleSheetsOutputVendor];
+const outputVendors = [ynabOutputVendor, googleSheetsOutputVendor, json];
 
 export default outputVendors;
