+++ src/originalBudgetTrackingApp/import/importTransactions.test.ts
@@ -1,5 +1,5 @@
+import { calculateTransactionHash } from '@/originalBudgetTrackingApp/transactions/transactions';
 import { TransactionStatuses, TransactionTypes } from '@brafdlog/israeli-bank-scrapers-core/lib/transactions';
-import { calculateTransactionHash } from './importTransactions';
 import { Transaction } from './bankScraper';
 
 const TRANSACTION_1: Transaction = {
